Homework #5 - Auto logging - Исправлены недочёты.
