+++ HW05-auto-logging/src/main/java/annotation/Log.java
@@ -1,3 +1,5 @@
+package annotation;
+
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
