+++ HW05-auto-logging/src/main/java/proxy/Main.java
@@ -1,7 +1,9 @@
+package proxy;
+
 public class Main {
 
     public static void main(String[] args) {
-        TestLoggingInterface myClass = LoggingProxy.createClass();
+        TestLoggingInterface myClass = LoggingProxy.createClass(new TestLogging());
         myClass.calculation(1);
         myClass.calculation(1, 2);
         myClass.calculation(1, 2, 3);
