+++ HW05-auto-logging/src/main/java/proxy/TestLogging.java
@@ -1,3 +1,7 @@
+package proxy;
+
+import annotation.Log;
+
 public class TestLogging implements TestLoggingInterface {
 
     @Log
