+++ HW05-auto-logging/src/main/java/proxy/TestLoggingInterface.java
@@ -1,3 +1,5 @@
+package proxy;
+
 public interface TestLoggingInterface {
 
     void calculation(int param);
